Implement Long-Short Term Memory Model with Pytorch
LSTM model is used to predict gauges measurement (time series) using
other sensors measurement.

The preprocessing phase consists of:
 - filling missing measurement using back fill method and linear
 interpolation
 - merging redundant sensors using truncated mean to void the effect
 of outliers.
 - creating derivative feature for each remaining input
 - adding history of all obtained features

For each epoch in the training phase, we choose a different order of
browsing flights. We also concatenate the samples of many flights and
we shuffle them to add stochasticity. Then, we divide data into batch
and we train the LSTM model on each batch.